# docs/swagger.yaml
definitions:
  handler.ErrorResponse:
    properties:
      errorCode:
        type: string
      message:
        type: string
    type: object
  handler.RegisterUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password_hash:
        type: string
    type: object
  handler.RegisterUserResponse:
    properties:
      data:
        $ref: '#/definitions/schemas.UserRegisterResponse'
      message:
        type: string
    type: object
  handler.ShowUserResponse:
    properties:
      data:
        $ref: '#/definitions/schemas.UserRegisterResponse'
      message:
        type: string
    type: object
  schemas.UserRegisterResponse:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password_hash:
        type: string
      updatedAt:
        type: string
    type: object
info:
  contact: {}
paths:
  /register:
    get:
      consumes:
      - application/json
      description: Show a user register
      parameters:
      - description: User identification
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ShowUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Show user
      tags:
      - Register
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RegisterUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Register user
      tags:
      - Register
swagger: "2.0"